+++ R/cf-workers.R
@@ -31,7 +31,9 @@ cf_list_workers_scripts <- function(
     token = token,
     email = email,
     api_key = api_key
-  )
+  ) |>
+    httr2::req_perform() |>
+    cf_resp()
   if (as_df) cf_records_to_df(records) else records
 }
 
@@ -68,7 +70,9 @@ cf_get_workers_script <- function(
     token = token,
     email = email,
     api_key = api_key
-  )
+  ) |>
+    httr2::req_perform() |>
+    cf_resp()
 }
 
 #' Workers invocations over time